Draw straight lines without antialias
Drawing lines with antialias is more difficult because all coordinates
must be exact and are influenced by the strokeWidth.

Furthermore, antialias drawing differs on different devices. My Acer
tablet draws them as expected, but both phones (Moto 4G, Samsung J5)
spill some of the paint to the bottom right so that the lines are not
as crispy as possible.

When drawing lines without antialias, the coordinates should always be
given as integer values. When giving coordinates of the form (n + 0.5),
the half-even rounding mode is used and produces shifted lines. This
rounding mode is also used for the strokeWidth, which makes the width
0.5 equivalent to 0.0, and the width 1.5 equivalent to 2.0.