Only show the guiding lines when making a move
When making a move, it is crucial that the move is placed at the correct
intersection, and that action cannot be undone. Therefore the guiding
lines are necessary there.

When editing the board and placing additional stones, there is no time
pressure, no immediate action, and everything can be undone manually.
Therefore, no guiding lines are necessary.