Add test for almost ko
This test prevents a probable mistake when programming the ko rule. The
rule is: If a single stone has captured a single stone, and this single
stone is captured back in the directly following move, it is not allowed
because of the ko rule.

In this situation, b1 is part of a larger chain.